requirements.txt
torch>=2.1.0
transformers>=4.57.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

python-docx>=1.1.0
lxml>=5.0.0

protonx

# GUI Desktop App
PyQt5>=5.15.0
